move optional_caster to detail/nb_optional.h
